% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{cardargus-package}
\alias{cardargus}
\alias{cardargus-package}
\title{cardargus: Generate SVG Information Cards with Embedded Fonts and Badges}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Create self-contained SVG information cards with embedded 'Google Fonts', shields-style badges, and custom logos. Cards are fully portable SVG files ideal for dashboards, reports, and web applications. Includes functions to export cards to PNG and PDF formats and display them in 'R Markdown' and 'Quarto' documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://strategicprojects.github.io/cardargus/}
  \item \url{https://github.com/StrategicProjects/cardargus}
  \item Report bugs at \url{https://github.com/StrategicProjects/cardargus/issues}
}

}
\author{
\strong{Maintainer}: Andre Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item Andre Leite \email{leite@castlab.org}
  \item Marcos Wasiliew \email{marcos.wasiliew@gmail.com} (\href{https://orcid.org/0009-0004-4694-3159}{ORCID})
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br}
  \item Carlos Amorim \email{carlos.agaf@ufpe.br}
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br}
  \item Júlia Nascimento Barreto \email{juliabarreto@gd.seplag.pe.gov.br} (\href{https://orcid.org/0009-0004-2851-7770}{ORCID})
}

}
\keyword{internal}
