% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{close_chrome_session}
\alias{close_chrome_session}
\title{Close the persistent Chrome session (if any)}
\usage{
close_chrome_session()
}
\value{
NULL (invisibly).
}
\description{
Close the persistent Chrome session (if any)
}
\keyword{internal}
