% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compress_number}
\alias{compress_number}
\title{Compress number to abbreviated format}
\usage{
compress_number(x, digits = 1)
}
\arguments{
\item{x}{Numeric value to compress}

\item{digits}{Number of decimal places}
}
\value{
Character string with abbreviated number
}
\description{
Compress number to abbreviated format
}
\examples{
compress_number(1234567)
compress_number(36400000)
}
