% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{css_gradient_to_svg}
\alias{css_gradient_to_svg}
\title{Convert CSS gradient to SVG gradient}
\usage{
css_gradient_to_svg(css_gradient, id = "grad")
}
\arguments{
\item{css_gradient}{CSS gradient string}

\item{id}{ID for the gradient}
}
\value{
SVG gradient definition
}
\description{
Convert CSS gradient to SVG gradient
}
\keyword{internal}
