% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.R
\name{custom_logo_svg}
\alias{custom_logo_svg}
\title{Generic logo placeholder SVG}
\usage{
custom_logo_svg(svg_content, width = 40, height = 40)
}
\arguments{
\item{svg_content}{Raw SVG content or path to SVG file}

\item{width}{Desired width}

\item{height}{Desired height}
}
\value{
SVG string
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=load_svg_for_embed]{load_svg_for_embed()}} instead,
or simply pass the SVG file path directly to \code{svg_card()}.
}
\keyword{internal}
