% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{detect_svg_fonts}
\alias{detect_svg_fonts}
\title{Detect font families used in an SVG}
\usage{
detect_svg_fonts(svg_content)
}
\arguments{
\item{svg_content}{Character SVG.}
}
\value{
Character vector of font family names (unique).
}
\description{
Detect font families used in an SVG
}
\keyword{internal}
