% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{download_google_font_woff2}
\alias{download_google_font_woff2}
\title{Download a Google Font as WOFF2 to the cache}
\usage{
download_google_font_woff2(family, weight = "400")
}
\arguments{
\item{family}{Font family name.}

\item{weight}{Numeric/character weight (default "400").}
}
\value{
Path to cached WOFF2 file or NA_character_.
}
\description{
Download a Google Font as WOFF2 to the cache
}
\keyword{internal}
