% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{embed_svg_fonts}
\alias{embed_svg_fonts}
\title{Embed a font file into an SVG via @font-face}
\usage{
embed_svg_fonts(svg_content, font_family, woff2_path)
}
\arguments{
\item{svg_content}{Character SVG.}

\item{font_family}{Font family name to embed.}

\item{woff2_path}{Path to a font file (.woff2, .ttf, .woff, or .otf).}
}
\value{
SVG string with embedded @font-face rule.
}
\description{
Embed a font file into an SVG via @font-face
}
\keyword{internal}
