% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ensure_output_dir}
\alias{ensure_output_dir}
\title{Ensure output directory exists}
\usage{
ensure_output_dir(path)
}
\arguments{
\item{path}{Output file path.}
}
\description{
Ensure output directory exists
}
\keyword{internal}
