% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_xml}
\alias{escape_xml}
\title{Escape XML special characters}
\usage{
escape_xml(text)
}
\arguments{
\item{text}{Text to escape}
}
\value{
Escaped text
}
\description{
Escape XML special characters
}
\keyword{internal}
