% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{font_available}
\alias{font_available}
\title{Check if a font is available for embedding}
\usage{
font_available(family)
}
\arguments{
\item{family}{Font family name}
}
\value{
TRUE if font is cached (any supported format), FALSE otherwise
}
\description{
Check if a font is available for embedding
}
\examples{
font_available("Jost")
font_available("Montserrat")
}
