% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{font_cache_path}
\alias{font_cache_path}
\title{Build a cache filename for a font family}
\usage{
font_cache_path(family)
}
\arguments{
\item{family}{Font family name.}
}
\value{
Path to the cached font file.
}
\description{
Build a cache filename for a font family
}
\keyword{internal}
