% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{fonts_dir}
\alias{fonts_dir}
\title{Get the path to fonts directory}
\usage{
fonts_dir()
}
\value{
Character string with the path to font cache directory
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=font_cache_dir]{font_cache_dir()}} instead to get the
directory where fonts are cached.
}
\keyword{internal}
