% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{get_chrome_download_info}
\alias{get_chrome_download_info}
\title{Get Chrome for Testing download info}
\usage{
get_chrome_download_info()
}
\description{
Get Chrome for Testing download info
}
\keyword{internal}
