% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{get_chrome_platform}
\alias{get_chrome_platform}
\title{Get platform identifier for Chrome downloads}
\usage{
get_chrome_platform()
}
\description{
Get platform identifier for Chrome downloads
}
\keyword{internal}
