% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_font_css}
\alias{get_font_css}
\title{Get Google Font CSS for embedding in SVG}
\usage{
get_font_css(font_name = "Jost", weights = c(400, 500, 600, 700))
}
\arguments{
\item{font_name}{Name of the Google Font}

\item{weights}{Vector of font weights to include}
}
\value{
Character string with CSS @font-face rules
}
\description{
Get Google Font CSS for embedding in SVG
}
\examples{
get_font_css("Jost")
}
