% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/logos.R
\name{building_icon_svg}
\alias{building_icon_svg}
\alias{icon_building}
\title{Building Icon SVG}
\usage{
icon_building(
  width = 50,
  height = 56,
  stroke_color = "white",
  stroke_width = 2,
  fill = "none"
)
}
\arguments{
\item{width}{Width of the icon}

\item{height}{Height of the icon}

\item{stroke_color}{Stroke color}

\item{stroke_width}{Stroke width}

\item{fill}{Fill color}
}
\value{
SVG string
}
\description{
Generate a building/apartment icon SVG.
}
\examples{
icon_building(50, 56)
}
