% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.R
\name{icon_construction}
\alias{icon_construction}
\title{Construction Icon SVG}
\usage{
icon_construction(
  width = 50,
  height = 56,
  stroke_color = "white",
  stroke_width = 2,
  fill = "none"
)
}
\arguments{
\item{width}{Width of the icon}

\item{height}{Height of the icon}

\item{stroke_color}{Stroke color}

\item{stroke_width}{Stroke width}

\item{fill}{Fill color}
}
\value{
SVG string
}
\description{
Generate a construction/crane icon SVG.
}
\examples{
icon_construction(50, 56)
}
