% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.R
\name{icon_house}
\alias{icon_house}
\title{House Icon SVG}
\usage{
icon_house(
  width = 50,
  height = 56,
  stroke_color = "white",
  stroke_width = 35,
  fill = "none"
)
}
\arguments{
\item{width}{Width of the icon}

\item{height}{Height of the icon}

\item{stroke_color}{Stroke color}

\item{stroke_width}{Stroke width}

\item{fill}{Fill color (default none)}
}
\value{
SVG string
}
\description{
Generate a house/home icon SVG. You can also use any SVG file path
instead of built-in icons.
}
\examples{
icon_house(50, 56)

# You can also use a custom SVG file:
# svg_card(..., with_icon = "/path/to/my_icon.svg")
}
