% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.R
\name{icon_map_pin}
\alias{icon_map_pin}
\title{Map Pin Icon SVG}
\usage{
icon_map_pin(
  width = 50,
  height = 56,
  stroke_color = "white",
  stroke_width = 2,
  fill = "none"
)
}
\arguments{
\item{width}{Width of the icon}

\item{height}{Height of the icon}

\item{stroke_color}{Stroke color}

\item{stroke_width}{Stroke width}

\item{fill}{Fill color}
}
\value{
SVG string
}
\description{
Generate a map pin/location icon SVG.
}
\examples{
icon_map_pin(50, 56)
}
