% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.R
\name{icon_money}
\alias{icon_money}
\title{Dollar/Money Icon SVG}
\usage{
icon_money(
  width = 50,
  height = 56,
  stroke_color = "white",
  stroke_width = 2,
  fill = "none"
)
}
\arguments{
\item{width}{Width of the icon}

\item{height}{Height of the icon}

\item{stroke_color}{Stroke color}

\item{stroke_width}{Stroke width}

\item{fill}{Fill color}
}
\value{
SVG string
}
\description{
Generate a dollar/money icon SVG.
}
\examples{
icon_money(50, 56)
}
