% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_light_color}
\alias{is_light_color}
\title{Check if a color is light}
\usage{
is_light_color(color)
}
\arguments{
\item{color}{A hex color string or color name}
}
\value{
Logical indicating if the color is light
}
\description{
Check if a color is light
}
\examples{
is_light_color("#FFFFFF")
is_light_color("#000000")
}
