% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{list_bundled_svgs}
\alias{list_bundled_svgs}
\title{List available bundled SVG files}
\usage{
list_bundled_svgs()
}
\value{
Character vector of available SVG filenames
}
\description{
List available bundled SVG files
}
