% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{list_fonts}
\alias{list_fonts}
\title{List registered/cached fonts}
\usage{
list_fonts()
}
\value{
Character vector of font family names that are cached
}
\description{
List registered/cached fonts
}
\examples{
list_fonts()
}
