% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{parse_viewbox}
\alias{parse_viewbox}
\title{Parse a viewBox string into width/height}
\usage{
parse_viewbox(vb)
}
\arguments{
\item{vb}{A viewBox string \code{"minx miny width height"} or \code{NA}.}
}
\value{
Numeric vector \code{c(width, height)}, possibly \code{c(NA, NA)}.
}
\description{
Parse a viewBox string into width/height
}
\keyword{internal}
