% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{prepare_svg_for_raster}
\alias{prepare_svg_for_raster}
\title{Prepare SVG for rasterization (sanitize + embed fonts)}
\usage{
prepare_svg_for_raster(svg_content)
}
\arguments{
\item{svg_content}{Character SVG.}
}
\value{
A sanitized SVG with embedded WOFF2 fonts.
}
\description{
Removes Google Fonts @import rules (not supported by librsvg), sanitizes
editor metadata (handled by sanitize_svg_for_raster()), then embeds fonts
referenced by the SVG into @font-face blocks.
}
\keyword{internal}
