% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_google_font}
\alias{register_google_font}
\title{Register Google Font (sysfonts)}
\usage{
register_google_font(font_family)
}
\arguments{
\item{font_family}{Font family name (e.g., "Jost")}
}
\value{
Invisible NULL
}
\description{
Registers a Google Font using \pkg{sysfonts}.
}
