% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{svg_dims}
\alias{svg_dims}
\title{Read SVG width and height robustly}
\usage{
svg_dims(svg_content, prefer = c("attr", "viewBox"))
}
\arguments{
\item{svg_content}{SVG string.}

\item{prefer}{Which source to try first per dimension (see \code{\link[=svg_dimension]{svg_dimension()}}).}
}
\value{
A list with numeric \code{width} and \code{height} (either may be NA).
}
\description{
Read SVG width and height robustly
}
\keyword{internal}
