% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{svg_font_families}
\alias{svg_font_families}
\title{Detect font families referenced by an SVG}
\usage{
svg_font_families(svg_content)
}
\arguments{
\item{svg_content}{SVG string.}
}
\value{
Unique character vector of font family names (generics removed).
}
\description{
Detect font families referenced by an SVG
}
\keyword{internal}
