% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{svg_font_families_regex}
\alias{svg_font_families_regex}
\title{Regex fallback for \code{\link[=svg_font_families]{svg_font_families()}}}
\usage{
svg_font_families_regex(svg_content)
}
\arguments{
\item{svg_content}{SVG string.}
}
\description{
Regex fallback for \code{\link[=svg_font_families]{svg_font_families()}}
}
\keyword{internal}
