% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{svg_geometry_regex}
\alias{svg_geometry_regex}
\title{Regex fallback for \code{\link[=svg_geometry]{svg_geometry()}}}
\usage{
svg_geometry_regex(svg_content)
}
\arguments{
\item{svg_content}{SVG string.}
}
\description{
Regex fallback for \code{\link[=svg_geometry]{svg_geometry()}}
}
\keyword{internal}
