% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{svgs_dir}
\alias{svgs_dir}
\title{Get the path to package SVGs directory}
\usage{
svgs_dir()
}
\value{
Character string with the path to SVGs directory
}
\description{
Returns the path to the inst/svgs directory where SVG files are stored.
}
