% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{text_height}
\alias{text_height}
\title{Calculate text height using gdtools}
\usage{
text_height(text, font = "Jost", fontsize = 12)
}
\arguments{
\item{text}{Text to measure}

\item{font}{Font name}

\item{fontsize}{Font size in pixels}
}
\value{
Numeric height in pixels
}
\description{
Calculate text height using gdtools
}
\keyword{internal}
