% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{text_width}
\alias{text_width}
\title{Calculate text width using gdtools}
\usage{
text_width(text, font = "Jost", fontsize = 12)
}
\arguments{
\item{text}{Text to measure}

\item{font}{Font name}

\item{fontsize}{Font size in pixels}
}
\value{
Numeric width in pixels
}
\description{
Calculate text width using gdtools
}
\keyword{internal}
