% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap text to fit within a given width}
\usage{
wrap_text(text, max_width, font = "Jost", fontsize = 12)
}
\arguments{
\item{text}{Text to wrap}

\item{max_width}{Maximum width in pixels}

\item{font}{Font name}

\item{fontsize}{Font size}
}
\value{
List with wrapped lines and total height
}
\description{
Wrap text to fit within a given width
}
\keyword{internal}
