% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{write_svg_tempfile}
\alias{write_svg_tempfile}
\title{Write SVG content to a temporary .svg file}
\usage{
write_svg_tempfile(svg_content)
}
\arguments{
\item{svg_content}{Character SVG content.}
}
\value{
Path to a temporary SVG file.
}
\description{
Write SVG content to a temporary .svg file
}
\keyword{internal}
