% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{cardargus-package}
\alias{cardargus}
\alias{cardargus-package}
\title{cardargus: Generate SVG Information Cards with Embedded Fonts and Badges}
\description{
Create self-contained SVG information cards with embedded 'Google Fonts', shields-style badges, and custom logos. Cards are fully portable SVG files ideal for dashboards, reports, and web applications. Includes functions to export cards to PNG format and display them in 'R Markdown' and 'Quarto' documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/cardargus}
  \item Report bugs at \url{https://github.com/StrategicProjects/cardargus/issues}
}

}
\author{
\strong{Maintainer}: Andre Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br}
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br}
}

}
\keyword{internal}
