% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/fonts.R
\name{cardargus_font_cache_dir}
\alias{cardargus_font_cache_dir}
\alias{font_cache_dir}
\title{Get font cache directory}
\usage{
font_cache_dir()
}
\value{
A character path to the cache directory.
}
\description{
Returns the directory where cardargus caches downloaded font files.
Fonts in this directory are automatically embedded in SVG/PNG exports.
}
\examples{
font_cache_dir()
# List cached fonts
list.files(font_cache_dir())
}
