% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_id}
\alias{generate_id}
\title{Generate unique ID suffix}
\usage{
generate_id(...)
}
\arguments{
\item{...}{Values to hash}
}
\value{
Character string with unique ID
}
\description{
Generate unique ID suffix
}
\keyword{internal}
